ex-elasticache/FlaskApp/requirements.txt
Flask
boto3
python-memcached
